# SDK/swaggerdoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Microsoft Non Fungible Token Service API Endpoint",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/ServiceManagement/BlockchainNetworks/BlockchainNetwork/{Id}": {
      "get": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "GetBlockchainNetworkById",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "UnRegisterBlockchainNetwork",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/ServiceManagement/BlockchainNetworks/BlockchainNetwork": {
      "post": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "RegisterBlockchainNetwork",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/BlockchainNetworkInfo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockchainNetworkInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockchainNetworkInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BlockchainNetworkInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockchainNetwork"
                }
              }
            }
          }
        }
      }
    },
    "/ServiceManagement/Parties/Party/{Id}": {
      "get": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "GetPartyById",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "UnRegisterParty",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/ServiceManagement/Parties/Party": {
      "post": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "RegisterParty",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PartyInfo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PartyInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Party"
                }
              }
            }
          }
        }
      }
    },
    "/ServiceManagement/Users/User/{Id}": {
      "get": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "GetUserById",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "UnRegisterUser",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/ServiceManagement/Users/User": {
      "get": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "GetUserByPublicAddress",
        "parameters": [
          {
            "name": "PublicAddress",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceManagement"
        ],
        "operationId": "RegisterUser",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockchainNetwork": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "blockchainPlatformName": {
            "type": "string",
            "nullable": true
          },
          "blockchainPlatformType": {
            "type": "string",
            "nullable": true
          },
          "blockchainNode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "__partitionkey": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BlockchainNetworkInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "nodeURL": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Party": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "partyName": {
            "type": "string",
            "nullable": true
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "__partitionkey": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PartyInfo": {
        "type": "object",
        "properties": {
          "partyName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "publicAddress": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "party": {
            "$ref": "#/components/schemas/Party"
          },
          "blockchainNetwork": {
            "$ref": "#/components/schemas/BlockchainNetwork"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "__partitionkey": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "partyID": {
            "type": "string",
            "format": "uuid"
          },
          "blockchainNetworkID": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}